requirements.txt
requests==2.25.1
discord-py-interactions==4.2.1